DEVX-2476: Add OW sequences support to the plugin
