+++ package.json
@@ -34,10 +34,11 @@
     "eslint-plugin-promise": "^5",
     "eslint-plugin-standard": "^4.0.1",
     "execa": "^4",
-    "jest": "^24",
+    "jest": "^25",
     "jest-haste-map": "^24.5.0",
     "jest-junit": "^6.0.0",
     "jest-resolve": "^24.5.0",
+    "jest-when": "^3.5.1",
     "stdout-stderr": "^0.1.9"
   },
   "engines": {
